requirements.txt
blinker==1.4
click==6.6
coverage==4.2
ecdsa==0.13
Fabric3==1.12.post1
Flask==0.11.1
Flask-Script==2.0.5
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.2
pandas==0.19.1
paramiko==1.17.2
pycrypto==2.6.1
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
Werkzeug==0.11.11
